+++ src/main/java/edu/eci/cvds/sampleprj/dao/mybatis/MyBATISClienteDAO.java
@@ -68,9 +68,9 @@ public class MyBATISClienteDAO implements ClienteDAO{
 			throw new PersistenceException("Error al consultar multa",e);
 		}
 	}
-        
+
     @Override
-    public long agregarItemRentado(int id,int idit, Date fechainicio,Date fechafinn) throws PersistenceException {
+    public long agregarItemRentado(long id, int idit, Date fechainicio, Date fechafinn) throws PersistenceException {
         try {
             return clienteMapper.agregarItemRentadoACliente(id,idit,fechainicio,fechafinn);
         }catch(org.apache.ibatis.exceptions.PersistenceException e) {
